+++ TorSharp/TorrentFile/Models/Content/File/EmptyFile.cs
@@ -2,6 +2,6 @@ namespace TorSharp.TorrentFile.Models.Content;
 
 public record EmptyFile(string Name) : IFile
 {
-    public int Length => 0;
+    public long Length => 0;
     public string? PiecesRoot => null;
 }
