+++ TorSharp/TorrentFile/Models/Content/File/File.cs
@@ -2,6 +2,6 @@ namespace TorSharp.TorrentFile.Models.Content;
 
 public record File(
     string Name,
-    int Length,
+    long Length,
     string PiecesRoot // It can only be nullable for the fact that an empty file does not contain a root (since there is no data).
 ) : IFile;
