+++ TorSharp/TorrentFile/Models/Info/IInfo.cs
@@ -5,7 +5,7 @@ namespace TorSharp.TorrentFile.Models.Info;
 public interface IInfo
 {
     public string TorrentName { get; }
-    public int PieceBytes { get; }
-    public int MetaVersion { get; }
+    public long PieceBytes { get; }
+    public long MetaVersion { get; }
     public IFileTree FileTree { get; }
 }
