+++ TorSharp/TorrentFile/Models/Info/Info.cs
@@ -2,4 +2,4 @@ using TorSharp.TorrentFile.Models.FileTree;
 
 namespace TorSharp.TorrentFile.Models.Info;
 
-public record Info(string TorrentName, int PieceBytes, int MetaVersion, IFileTree FileTree) : IInfo;
+public record Info(string TorrentName, long PieceBytes, long MetaVersion, IFileTree FileTree) : IInfo;
